[TD-235] 타입 관련 클래스들 연관된 도메인 패키지에 위치시킴
