+++ threedollar-api/src/main/java/com/depromeet/threedollar/api/controller/advice/ServerExceptionEventListener.java
@@ -1,6 +1,6 @@
 package com.depromeet.threedollar.api.controller.advice;
 
-import com.depromeet.threedollar.api.event.ServerExceptionOccurredEvent;
+import com.depromeet.threedollar.api.event.ServerErrorOccuredEvent;
 import com.depromeet.threedollar.external.client.slack.SlackApiClient;
 import com.depromeet.threedollar.external.client.slack.dto.request.PostSlackMessageRequest;
 import io.sentry.Sentry;
@@ -17,7 +17,7 @@ public class ServerExceptionEventListener {
     private final SlackApiClient slackApiCaller;
 
     @EventListener
-    public void sendErrorNotification(ServerExceptionOccurredEvent event) {
+    public void sendErrorNotification(ServerErrorOccuredEvent event) {
         Sentry.captureException(event.getException());
 
         slackApiCaller.postMessage(PostSlackMessageRequest.of(String.format(ERROR_MESSAGE_FORMAT,
