+++ threedollar-api/src/main/java/com/depromeet/threedollar/api/event/ServerErrorOccuredEvent.java
@@ -1,7 +1,7 @@
 package com.depromeet.threedollar.api.event;
 
 import com.depromeet.threedollar.common.exception.ErrorCode;
-import com.depromeet.threedollar.domain.type.ServerEventType;
+import com.depromeet.threedollar.domain.domain.common.ServerEventType;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 
 @Getter
 @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
-public class ServerExceptionOccurredEvent {
+public class ServerErrorOccuredEvent {
 
     private final ServerEventType type;
 
@@ -20,8 +20,8 @@ public class ServerExceptionOccurredEvent {
 
     private final LocalDateTime timeStamp;
 
-    public static ServerExceptionOccurredEvent error(ErrorCode errorCode, Exception exception, LocalDateTime timeStamp) {
-        return new ServerExceptionOccurredEvent(ServerEventType.ERROR, errorCode, exception, timeStamp);
+    public static ServerErrorOccuredEvent error(ErrorCode errorCode, Exception exception, LocalDateTime timeStamp) {
+        return new ServerErrorOccuredEvent(ServerEventType.ERROR, errorCode, exception, timeStamp);
     }
 
 }
