+++ threedollar-application-service/src/main/java/com/depromeet/threedollar/application/mapper/faq/FaqCategoryMapper.kt
@@ -1,6 +1,6 @@
-package com.depromeet.threedollar.application.mapper
+package com.depromeet.threedollar.application.mapper.faq
 
-import com.depromeet.threedollar.application.mapper.dto.FaqCategoryResponse
+import com.depromeet.threedollar.application.mapper.faq.dto.response.FaqCategoryResponse
 import com.depromeet.threedollar.domain.domain.faq.FaqCategory
 
 object FaqCategoryMapper {
