+++ threedollar-application-service/src/main/java/com/depromeet/threedollar/application/mapper/faq/dto/response/FaqCategoryResponse.kt
@@ -1,0 +1,17 @@
+package com.depromeet.threedollar.application.mapper.faq.dto.response
+
+import com.depromeet.threedollar.domain.domain.faq.FaqCategory
+
+data class FaqCategoryResponse(
+    val category: FaqCategory,
+    val description: String,
+    val displayOrder: Int
+) {
+
+    companion object {
+        fun of(category: FaqCategory): FaqCategoryResponse {
+            return FaqCategoryResponse(category, category.description, category.displayOrder)
+        }
+    }
+
+}
