+++ threedollar-application-service/src/main/java/com/depromeet/threedollar/application/mapper/menu/dto/response/MenuCategoryResponse.kt
@@ -1,23 +1,7 @@
-package com.depromeet.threedollar.application.mapper.dto
+package com.depromeet.threedollar.application.mapper.menu.dto.response
 
-import com.depromeet.threedollar.domain.domain.faq.FaqCategory
 import com.depromeet.threedollar.domain.domain.menu.MenuCategoryType
 import com.fasterxml.jackson.annotation.JsonProperty
-
-
-data class FaqCategoryResponse(
-    val category: FaqCategory,
-    val description: String,
-    val displayOrder: Int
-) {
-
-    companion object {
-        fun of(category: FaqCategory): FaqCategoryResponse {
-            return FaqCategoryResponse(category, category.description, category.displayOrder)
-        }
-    }
-
-}
 
 data class MenuCategoryResponse(
     val category: MenuCategoryType,
