+++ threedollar-domain/src/main/java/com/depromeet/threedollar/domain/domain/common/DistanceGroupType.java
@@ -1,4 +1,4 @@
-package com.depromeet.threedollar.domain.type;
+package com.depromeet.threedollar.domain.domain.common;
 
 import com.depromeet.threedollar.common.exception.model.InternalServerException;
 import lombok.AccessLevel;
