+++ threedollar-domain/src/main/java/com/depromeet/threedollar/domain/domain/common/ServerEventType.java
@@ -1,0 +1,8 @@
+package com.depromeet.threedollar.domain.domain.common;
+
+public enum ServerEventType {
+
+    INFO,
+    ERROR
+
+}
