--- threedollar-domain/src/main/java/com/depromeet/threedollar/domain/type/ServerEventType.java
@@ -1,8 +1,0 @@
-package com.depromeet.threedollar.domain.type;
-
-public enum ServerEventType {
-
-    INFO,
-    ERROR
-
-}
