+++ threedollar-domain/src/test/java/com/depromeet/threedollar/domain/domain/review/RatingGroupTypeTest.java
@@ -1,4 +1,4 @@
-package com.depromeet.threedollar.domain.type;
+package com.depromeet.threedollar.domain.domain.review;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
